Fixed the Random Quote function
