+++ server/server.js
@@ -32,7 +32,7 @@ app.get(`/api/quotes/random`, (req, res) => {
   })
 })
 
-app.get(`/api/search/quotes`, (req, res) => {
+app.get(`/api/search/quotes/`, (req, res) => {
   axios.get(`https://api.quotable.io/search/quotes?query=${req.params.id}`).then((response) => {
     console.log('Checking the response', response.data)
     res.send(response.data);
