+++ src/redux/sagas/quote.saga.js
@@ -6,7 +6,7 @@ function* getQuote(action){
         console.log('Checking payload', action.payload);
         const quotes = yield axios.get(`/api/search/quotes?query=${action.payload}`);
         console.log('Checking quotes.data', quotes.data)
-        yield put({ type: 'SET_QUOTES', payload: quotes.data });
+        yield put({ type: 'SET_QUOTES', payload: quotes.data.data });
 
     } catch (error){
         console.log('Error in getQuote', error);
